Agregue autenticacion de urls para diferentes tipos de usuarios, Ademas ahora se pueden filtrar los comunicados que reciben los padres

+++ villadaApp/app/views.py
@@ -22,11 +22,10 @@ from rest_framework.permissions import IsAdminUser
 from rest_framework.authtoken.views import *
 from .models import CustomUser as User
 
-def index(request):
-    return render(request,'index.html')
-
-@login_required
+@login_required(login_url="/")
 def eliminarComunicados(request, id_comunicado):
+    if request.user.groups.filter(name='Padres').exists():
+        return redirect('comunicados_padres')
     comunicado = Comunicado.objects.get(id= id_comunicado)
     if request.method == "POST":
         comunicado.delete()
@@ -34,21 +33,11 @@ def eliminarComunicados(request, id_comunicado):
     return render(request, 'delete_comunicado.html',{'comunicado':comunicado})
 
 
-#Todavia no usamos esta vista, pero la dejo por si las dudas despues la usamos...
-@login_required
-def editarComunicados(request, id_comunicado):
-    comunicado = Comunicado.objects.get(id= id_comunicado)
-    if request.method == "GET":
-        form = ComunicadoForm(instance=comunicado)
-    else:
-        form = ComunicadoForm(request.POST, instance=comunicado)
-        if form.is_valid():
-            form.save
-        return redirect('comunicados')
-    return render(request, "redactar.html")
 
-@login_required
+@login_required(login_url="/")
 def redactar(request):
+    if request.user.groups.filter(name='Padres').exists():
+        return redirect('comunicados_padres')
     if request.method == "POST":
         form = ComunicadoForm(request.POST)
         if form.is_valid():
@@ -60,8 +49,10 @@ def redactar(request):
 
 
 
-@login_required
+@login_required(login_url="/")
 def comunicados(request):
+    if request.user.groups.filter(name='Padres').exists():
+        return redirect('comunicados_padres')
     queryset = request.GET.get("buscar")
     comunicados = Comunicado.objects.all().order_by('fecha')
     if queryset:
@@ -72,34 +63,31 @@ def comunicados(request):
 
     return render(request ,'comunicados.html', {'comunicados':comunicados})
 
-@login_required
+@login_required(login_url="/")
 def display_comunicado(request, id_comunicado):
     comunicado = Comunicado.objects.get(id= id_comunicado)
     return render(request, 'comunicado.html',{'comunicado':comunicado})
 
-@login_required
+@login_required(login_url="/")
 def ordenar_por_dir(request, order):
     comunicado = Comunicado.objects.all().order_by('directivo')
     return render(request, 'comunicado.html',{'comunicados':comunicados})
 
-@login_required
-def special(request):
-    return HttpResponse("You are logged in !")
 
-@login_required
+@login_required(login_url="/")
 def user_logout(request):
     logout(request)
     return HttpResponseRedirect(reverse('user_login'))
 
 def user_login(request):
-
+    logout(request)
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
         user = authenticate(username=username, password=password)
         if user is not None and user.groups.filter(name='Padres').exists():
             login(request, user)
-            return redirect('hola_padres')
+            return redirect('comunicados_padres')
         elif user is not None and user.is_superuser:
             login(request, user)
             return redirect('comunicados')
@@ -109,9 +97,6 @@ def user_login(request):
         return render(request, 'login.html', {})
 
 
-""" def curso_tipo(response):
-    curso = response.get["curso"]
-    return ComunicadoCurso.objects.all().filter(cursos_name = curso).values_list('id', flat=True) """
 
 class ComunicadoViewSet(viewsets.ModelViewSet):
     queryset = Comunicado.objects.all()
